requirements.txt
click==7.1.2
mysql-connector-python==8.0.20
protobuf==3.12.1
six==1.15.0
